backend/requirements.txt
fastapi[all]
aiosqlite
asyncpg
bcrypt==3.2.2
passlib[bcrypt]
python-dotenv
python-jose[cryptography]
sqlalchemy>=2.0.0
sqlalchemy[asyncio]